feat(CCGs): Migrate CCGs to Sub ICB Locations
The code for the Sub ICB Locations in England, LHBs in Wales, CHPs in Scotland, LCG in Northern Ireland and PHD in the Isle of Man; there are no equivalent areas in the Channel Islands (for which a pseudo code is included).

BREAKING CHANGE:

From July 2022, CCGs were replaced by Sub-ICB locations.